requirements.txt
flask
joblib
pandas
numpy
nltk
wordfreq
vaderSentiment
textblob
faiss-cpu
sentence_transformers
xgboost
scikit-learn==1.6.1